feat: investigate instruction date population in detail model
Look at the instruction request service when it populates the detail model, if the instruction-request.detail.instruction_date is empty, it should copy preferred_datetime or asynchronous_instruction_ready_date (this is a date, not a datetime so just set it to 8:00 AM). The goal here is when a new request is created, instruction date should be populated from the corresponding field based on the instruction_type. When a request is edited, the field value is updated. Also make sure instruction-request.duration is being copied to instruction-request.detail.instruction_duration

```git-revs
34a8de3  (Base revision)
a4f200b  Fix instruction date population to add 8:00 AM time for asynchronous dates and ensure duration is always copied to instruction_duration
HEAD     Set default duration of 30 minutes for asynchronous requests since they don't have duration field in form
```

codemcp-id: 140-feat-investigate-instruction-date-population-in-de