feat: investigate L-shaped navigation layout refactoring
investigate my project, with the goal of refactoring the navbar and layout into a more standard "L" navigation with a column on the left (currently navigation links are horizontal at the top), but still keep the user dropdown menu on the top right, and the dark mode toggle too. This is just an investigation, I'd like to know how much refactoring wouold be involved, what files would need to be changed. I do not want to change any of the functional elements of the app, just the presentation. I'm attaching a screenshot for reference. I do not want to duplicate the functionality of the screenshot, just the layout in general.

```git-revs
765e559  (Base revision)
HEAD     Create L-shaped dashboard mockup with sidebar navigation
```

codemcp-id: 153-feat-investigate-l-shaped-navigation-layout-refact